Separate the "project" lock into two separate locks, one that must be acquired by a non-isolated task that belongs to the project and one that must be acquired to access the mutable state of the project.
When parallel execution is enabled, these are the same. When it is not enabled, they are separate. This allows work from other builds in the tree to access the mutable state of a build while a task for that build is blocked while also preventing multiple tasks from the same build being started.